+++ TableView in CollectionView.xcodeproj/xcuserdata/hajimeito.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "EF73CD15-C803-4DEB-9070-A54432A878BF"
+   type = "1"
+   version = "2.0">
+</Bucket>
